finance-agent-llama-stack/requirements.txt
# Requirements for llama_stack_agent_responses.py
# Llama Stack agent responses standalone script

# Core dependencies
llama-stack-client==0.3.1
python-dotenv==1.2.1

# MCP Server dependencies
fastmcp
httpx
mcp

# Testing dependencies (optional)
pytest
pytest-asyncio
